requirements.txt
certifi==2017.11.5
chardet==3.0.4
dj-static==0.0.6
Django==1.11.6
django-ckeditor==5.3.1
django-colorfield==0.1.14
django-cors-headers==2.1.0
django-debug-toolbar==1.9.1
django-easy-select2==1.3.4
django-js-asset==0.1.1
django-material==1.1.1
django-rest-auth==0.9.2
django-smart-selects==1.5.2
django-widget-tweaks==1.4.1
djangorestframework==3.7.3
gunicorn==19.7.1
idna==2.6
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pystache==0.5.4
pytz==2017.3
redis==2.10.6
request==0.0.22
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
static==1.1.1
static3==0.7.0
urllib3==1.22
uWSGI==2.0.15
Willow==1.1
